+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Exclude deleted files from the paths passed to commands using `staged_only`.
+
 ## [1.8.0](https://github.com/moulibheemaneti/dart_husky/compare/dart_husky-v1.7.1...dart_husky-v1.8.0) (2026-08-21)